pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leadqualenv"
version = "2.0.0"
description = "Deterministic benchmark for outbound lead qualification in real estate sales."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["leadqualenv*", "server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
